Delete follows API convention instead.

+++ app/Http/Controllers/Api/UserController.php
@@ -7,14 +7,17 @@ use Illuminate\Http\Request;
 
 class UserController extends Controller
 {
-    public function selfDelete(Request $request)
+    public function delete(Request $request, int $id)
     {
         $user = $request->user();
+        if($user->id == $id){
+            // TODO: delete medias, playlists, etc.
+            $user->delete();
 
-        // TODO: delete medias, playlists, etc.
+            return response()->noContent();
+        }
+        else // TODO: admin deletion of another account
+            return response()->json(['message' => 'Not implemented'], 501);
 
-        $user->delete();
-
-        return response()->noContent();
     }
 }
